Refactor KeyOrValueSpecs into impl classes
